Simulation of HiC and WGS works

+++ src/svphase/utils/config.py
@@ -6,7 +6,7 @@ Created on Jan 29, 2014
 import numpy as np
 READLENGTH=100
 RAND_BUFFER_SIZE = 50000
-CONTIG_POS_TYPE = np.uint16
+CONTIG_POS_TYPE = np.uint32
 CONTIG_TYPE_MAX = np.iinfo(CONTIG_POS_TYPE).max
 ZONE_TYPE = np.uint8
 REFERENCE_HG18="/home/anand/data/hg/18/hg18.fa"
